backend/py-ai-api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
pydantic
pydantic[email]
# numpy==1.21.2 # Commented out, let pandas/sklearn guide this or install a specific version if needed later
pandas==1.3.3
scikit-learn==0.24.2
# tensorflow==2.6.0 # Removed as per user request
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5 
pyproj==3.6.1
shapely==2.0.7

# PyTorch (CPU version - check PyTorch website for specific versions/CUDA)
torch
torchvision
# torchaudio # Optional, uncomment if needed for audio processing